refactor common to integration
